Resolve Ming Feature dependencies
