+++ Stratis.Bitcoin/Features/Miner/MiningFeature.cs
@@ -1,12 +1,20 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Stratis.Bitcoin.Builder;
 using Stratis.Bitcoin.Builder.Feature;
+using Stratis.Bitcoin.Configuration.Logging;
 using Stratis.Bitcoin.Features.Wallet;
 
 namespace Stratis.Bitcoin.Features.Miner
 {
     public class MiningFeature : FullNodeFeature
     {
+        public MiningRPCController Controller { get; private set; }
+
+        public MiningFeature(MiningRPCController controller = null)
+        {
+            this.Controller = controller;
+        }
+
         ///<inheritdoc />
         public override void Start()
         {
@@ -34,6 +42,9 @@ namespace Stratis.Bitcoin.Features.Miner
     {
         public static IFullNodeBuilder AddMining(this IFullNodeBuilder fullNodeBuilder)
         {
+            LoggingConfiguration.RegisterLoggingConfiguration("mining", 
+                $"{ nameof(Stratis)}.{ nameof(Stratis.Bitcoin)}.{ nameof(Stratis.Bitcoin.Features)}.{ nameof(Stratis.Bitcoin.Features.Miner)}.*");
+
             fullNodeBuilder.ConfigureFeature(features =>
             {
                 features
@@ -42,9 +53,10 @@ namespace Stratis.Bitcoin.Features.Miner
                     {
                         services.AddSingleton<PowMining>();
                         services.AddSingleton<AssemblerFactory, PowAssemblerFactory>();
+                        services.AddSingleton<MiningRPCController>();
                     });
             });
-
+            
             return fullNodeBuilder;
         }
 
@@ -59,6 +71,7 @@ namespace Stratis.Bitcoin.Features.Miner
                         services.AddSingleton<PowMining>();
                         services.AddSingleton<PosMinting>();
                         services.AddSingleton<AssemblerFactory, PosAssemblerFactory>();
+                        services.AddSingleton<MiningRPCController>();
                     });
             });
