+++ Stratis.Bitcoin/Features/Miner/PowMining.cs
@@ -1,13 +1,10 @@
 ﻿using Microsoft.Extensions.Logging;
 using NBitcoin;
-using NBitcoin.BouncyCastle.Math;
 using Stratis.Bitcoin.Base;
-using Stratis.Bitcoin.Features.BlockStore;
 using Stratis.Bitcoin.Features.Consensus;
 using Stratis.Bitcoin.Utilities;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using IBlockRepository = Stratis.Bitcoin.Features.BlockStore.IBlockRepository;
 
@@ -193,66 +190,6 @@ namespace Stratis.Bitcoin.Features.Miner
 
             Guard.Assert(txCoinbase.Inputs[0].ScriptSig.Length <= 100);
             pblock.UpdateMerkleRoot();
-        }
-
-        public static Target GetWorkRequired(NBitcoin.Consensus consensus, ChainedBlock chainedBlock)
-        {
-            // Genesis block
-            if (chainedBlock.Height == 0)
-                return consensus.PowLimit;
-            var nProofOfWorkLimit = consensus.PowLimit;
-            var pindexLast = chainedBlock.Previous;
-            var height = chainedBlock.Height;
-
-            if (pindexLast == null)
-                return nProofOfWorkLimit;
-
-            // Only change once per interval
-            if ((height) % consensus.DifficultyAdjustmentInterval != 0)
-            {
-                if (consensus.PowAllowMinDifficultyBlocks)
-                {
-                    // Special difficulty rule for testnet:
-                    // If the new block's timestamp is more than 2* 10 minutes
-                    // then allow mining of a min-difficulty block.
-                    if (chainedBlock.Header.BlockTime > pindexLast.Header.BlockTime + TimeSpan.FromTicks(consensus.PowTargetSpacing.Ticks * 2))
-                        return nProofOfWorkLimit;
-                    else
-                    {
-                        // Return the last non-special-min-difficulty-rules-block
-                        ChainedBlock pindex = pindexLast;
-                        while (pindex.Previous != null && (pindex.Height % consensus.DifficultyAdjustmentInterval) != 0 && pindex.Header.Bits == nProofOfWorkLimit)
-                            pindex = pindex.Previous;
-                        return pindex.Header.Bits;
-                    }
-                }
-                return pindexLast.Header.Bits;
-            }
-
-            // Go back by what we want to be 14 days worth of blocks
-            var pastHeight = pindexLast.Height - (consensus.DifficultyAdjustmentInterval - 1);
-            ChainedBlock pindexFirst = chainedBlock.EnumerateToGenesis().FirstOrDefault(o => o.Height == pastHeight);
-            Guard.Assert(pindexFirst != null);
-
-            if (consensus.PowNoRetargeting)
-                return pindexLast.Header.Bits;
-
-            // Limit adjustment step
-            var nActualTimespan = pindexLast.Header.BlockTime - pindexFirst.Header.BlockTime;
-            if (nActualTimespan < TimeSpan.FromTicks(consensus.PowTargetTimespan.Ticks / 4))
-                nActualTimespan = TimeSpan.FromTicks(consensus.PowTargetTimespan.Ticks / 4);
-            if (nActualTimespan > TimeSpan.FromTicks(consensus.PowTargetTimespan.Ticks * 4))
-                nActualTimespan = TimeSpan.FromTicks(consensus.PowTargetTimespan.Ticks * 4);
-
-            // Retarget
-            var bnNew = pindexLast.Header.Bits.ToBigInteger();
-            bnNew = bnNew.Multiply(BigInteger.ValueOf((long)nActualTimespan.TotalSeconds));
-            bnNew = bnNew.Divide(BigInteger.ValueOf((long)consensus.PowTargetTimespan.TotalSeconds));
-            var newTarget = new Target(bnNew);
-            if (newTarget > nProofOfWorkLimit)
-                newTarget = nProofOfWorkLimit;
-
-            return newTarget;
         }
     }
 }
